pyproject.toml
[tool.poetry]
name = "bot2"
version = "0.1.0"
description = "A discord bot made with interactions.py."
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^0.20.0"
discord-py-interactions= "^5.8.0"
sentry-sdk = "^1.29.2"
motor = "^3.2.0"
pygithub = "^1.59.1"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
